отзывы - готово
Добавление, удаление, вывод отзывов в dashboard и в продукте

+++ backend/src/review/review.service.ts
@@ -1,7 +1,13 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadGatewayException,
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import CreateReviewDto from './dto/create-review.dto';
 import { ProductService } from 'src/product/product.service';
+import { ERole } from 'generated/prisma';
 
 @Injectable()
 export class ReviewService {
@@ -10,9 +16,32 @@ export class ReviewService {
     private productService: ProductService,
   ) {}
 
-  async getAll(userId: string) {
+  private async getOneById(id: string) {
+    const review = await this.prisma.review.findFirst({ where: { id } });
+    if (!review) throw new NotFoundException('Review is not found!');
+    return review;
+  }
+
+  async getAll() {
+    const reviews = await this.prisma.review.findMany({
+      include: { user: true },
+      orderBy: { createdAt: 'desc' },
+    });
+
+    return reviews;
+  }
+  async getAllByUserId(userId: string) {
     const reviews = await this.prisma.review.findMany({
       where: { userId },
+      include: { user: true },
+      orderBy: { createdAt: 'desc' },
+    });
+
+    return reviews;
+  }
+  async getAllByProductId(productId: string) {
+    const reviews = await this.prisma.review.findMany({
+      where: { productId },
       include: { user: true },
       orderBy: { createdAt: 'desc' },
     });
@@ -31,11 +60,11 @@ export class ReviewService {
     return reviews;
   }
 
-  async create(productId: string, dto: CreateReviewDto) {
+  async create(productId: string, userId: string, dto: CreateReviewDto) {
     const product = await this.productService.findById(productId);
-
+    console.log(dto);
     const review = await this.prisma.review.create({
-      data: { ...dto, productId: product.id },
+      data: { ...dto, userId, productId: product.id },
     });
 
     if (!review) throw new BadRequestException('Failed to create review');
@@ -54,11 +83,15 @@ export class ReviewService {
     return reviews;
   }
 
-  async delete(id: string) {
-    const review = await this.prisma.review.delete({ where: { id } });
+  async delete(id: string, userId: string, userRole: ERole) {
+    const review = await this.getOneById(id);
 
-    if (!review) throw new BadRequestException('Failed to delete review');
+    if (review.userId !== userId && userRole !== ERole.ADMIN)
+      throw new BadGatewayException("It isn't your review!");
+    const deleted = await this.prisma.review.delete({ where: { id } });
 
-    return review;
+    if (!deleted) throw new BadRequestException('Failed to delete review');
+
+    return deleted;
   }
 }
